Refactore & update
