+++ src/App.js
@@ -1,8 +1,32 @@
 import React from 'react';
-import RouterRoot from './router/Router';
+import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
+import DesignerPage from './pages/DesignerPage';
+import TablePage from './pages/table/TablePage';
 
-function App() {
-  return <RouterRoot />;
+export default function App() {
+  return (
+    <Router>
+      <Switch>
+        {ROUTES.map((route, i) => (
+          <Route
+            exact
+            key={i}
+            path={route.path}
+            render={() => <route.component />}
+          />
+        ))}
+      </Switch>
+    </Router>
+  );
 }
 
-export default App;
+const ROUTES = [
+  {
+    path: '/',
+    component: DesignerPage,
+  },
+  {
+    path: '/table',
+    component: TablePage,
+  },
+];
