--- src/pages/Pages.json
@@ -1,10 +1,0 @@
-{
-  "root": "/",
-  "login": "/login",
-  "table": "/table",
-  "createAccount": "/create-account",
-  "forgetPassword": "/forget-password",
-  "updatePassowrd": "/update-password",
-  "designer": "/designer",
-  "main": "/main"
-}
