+++ src/pages/designer/SVGRenderer.js
