+++ src/pages/designer/actions/Rotator.js
