+++ src/pages/designer/editors/BezierEditor.js
