+++ src/pages/designer/index.js
@@ -1,4 +1,4 @@
-export Preview from './Preview';
+export { Preview } from './Preview';
 export { Vector, Path, Rect, Circle, Text, Image } from './objects';
 export {
   TextPanel,
@@ -7,4 +7,4 @@ export {
   ArrangePanel,
   ImagePanel,
 } from './panels';
-export default from './Designer';
+export { default } from './Designer';
