+++ src/pages/designer/panels/ColorInput.js
