+++ src/pages/designer/panels/StylePanel.js
